python/requirements.txt
google-generativeai
python-dotenv
requests